requirements.txt
# RAG 系统依赖包

# LangChain 核心
langchain>=0.1.0
langchain-community>=0.0.10

# LLM SDK（支持多模型切换）
anthropic>=0.18.0
openai>=1.12.0

# 向量数据库
qdrant-client>=1.7.0

# 嵌入模型 & Reranker
sentence-transformers>=2.2.0
transformers>=4.36.0
torch>=2.0.0

# 文档处理
pypdf2>=3.0.0
python-docx>=1.0.0
markdown>=3.5.0
beautifulsoup4>=4.12.0

# Web 服务
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# 工具
python-dotenv>=1.0.0
tiktoken>=0.5.0
numpy>=1.24.0
pandas>=2.1.0

# MCP Server
mcp[cli]>=1.0.0

# 代码解析（可选，如需要请手动安装兼容版本）
# tree-sitter>=0.20.0
# tree-sitter-languages>=1.7.0
